fix: Configure git pull to rebase instead of merge
If the branch has not been updated while the action is running, then the
pull will be a no-op either way, and the push will succeed. If the
branch has been updated since the repository was initially checked out,
then rebase rather than merge when pulling so that no merge commit will
be created. If the rebase can not be performed non-interactively, then
it will fail immediately, and there's nothing further that can be done.
Configuring pull behavior has the added benefit of eliminating a
twelve-line warning from Git about pulling without specifying how to
reconcile divergent branches.